requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
google-genai>=0.2.0
python-multipart>=0.0.6
motor>=3.6.0
dnspython>=2.6.0
pymupdf>=1.23.0
passlib[argon2,bcrypt]>=1.7.4
pymongo[encryption]>=4.2
sse-starlette>=2.0.0
azure-storage-blob>=12.19.0
aiohttp>=3.9.0
PyJWT[crypto]>=2.8.0

# Testing Integration
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
